as.graphicsAnnot {grDevices}	R Documentation

Coerce an Object for Graphics Annotation

Description

Coerce an R object into a form suitable for graphics annotation.

Usage


as.graphicsAnnot(x)

Arguments

x
an R object

Details

Expressions, calls and names (as used by plotmath) are passed through unchanged. All other objects with an explicit class (as determined by is.object) are coerced by as.character to character vectors.

All the graphics and grid functions which use this coerce calls and names to expressions internally.

Value

A language object or a character vector.

[Package grDevices version 3.6.1 Index]