dev.xxx {grDevices}	R Documentation

Control Multiple Devices

Description

These functions provide control over multiple graphics devices.

Only one device is the active device. This is the device in which all graphics operations occur.

Devices are associated with a name (e.g., "X11" or "postscript") and a number; the "null device" is always device 1.

dev.off shuts down the specified (by default the current) device. graphics.off() shuts down all open graphics devices.

dev.set makes the specified device the active device.

Usage

dev.cur()
dev.list()
dev.next(which = dev.cur())
dev.prev(which = dev.cur())
dev.off(which = dev.cur())
dev.set(which = dev.next())
graphics.off()

Arguments

which 	 An integer specifying a device number

Value

dev.cur returns the number and name of the active device, or 1, the null device, if none is active.
dev.list returns the numbers of all open devices, except device 1, the null device. This is a numeric vector with a names attribute giving the names, or NULL is there is no open device.
dev.next and dev.prev return the number and name of the next / previous device in the list of devices. The list is regarded as a circular list, and "null device" will be included only if there are no open devices.
dev.off returns the name and number of the new active device (after the specified device has been shut down).
dev.set returns the name and number of the new active device.

See Also

Devices, such as postscript, etc.

layout and its links for setting up plotting regions on the current device.

Examples

## Not run: 
## Unix-specific example
x11()
plot(1:10)
x11()
plot(rnorm(10))
dev.set(dev.prev())
abline(0,1)# through the 1:10 points
dev.set(dev.next())
abline(h=0, col="gray")# for the residual plot
dev.set(dev.prev())
dev.off(); dev.off()#- close the two X devices
## End(Not run)

[Package grDevices version 2.1.0 Index]