dev.capture {grDevices}	R Documentation

Capture device output as a raster image

Description

dev.capture captures the current contents of a graphics device as a raster (bitmap) image.

Usage


dev.capture(native = FALSE)

Arguments

native
Logical. If FALSE the result is a matrix of R color names, if TRUE the output is returned as a nativeRaster object which is more efficient for plotting, but not portable.

Details

Not all devices support capture of the output as raster bitmaps. Typically, only image-based devices do and even not all of them.

Value

NULL if the device does not support capture, otherwise a matrix of color names (for native = FALSE) or a nativeRaster object (for native = TRUE).

[Package grDevices version 3.6.1 Index]