grDevices-defunct {grDevices}	R Documentation

Defunct Functions in Package grDevices

Description

The functions or variables listed here are no longer part of R as they are not needed (any more).

Usage

postscriptFont(family, metrics, encoding = "default")

Arguments

family 	 a character string giving the name of a Type 1 font family.

metrics 	 a vector of four or five strings giving paths to the afm (font metric) files for the font.

encoding 	 the name of an encoding file.

Details

The function postscriptFont has been replaced by Type1Font, which is just a change of name.

Value

One or more font mappings.

See Also

Defunct

[Package grDevices version 2.9.2 Index]