grSoftVersion {grDevices}	R Documentation

Report Versions of Graphics Software

Description

Report versions of third-party graphics software available on the current platform for R's graphics.

Usage


grSoftVersion()

Value

A named character vector containing at least the elements

cairo
the version of cairographics in use, or "" if cairographics is not available.

It may also contain the versions of third-party software used by the standard (on Windows), or X11-based (on Unix-alikes) bitmap devices:

libpng
the version of libpng in use, or "" if not available.

jpeg
the version of the JPEG headers used for compilation, or "" if JPEG support was not compiled in.

libtiff
the version of libtiff in use, or "" if not available.

It is conceivable but unlikely that the cairo-based bitmap devices will use different versions linked via cairographics, especially png(type = "cairo-png").

On macOS, if available, the Quartz-based devices will use the system versions of these libraries rather than those reported here.

Unless otherwise stated the reported version is that of the (possibly dynamically-linked) library in use at runtime.

Note that libjpeg-turbo as used on some Linux distributions reports its version as "6.2", the IJG version from which it forked.

See Also

extSoftVersion for versions of non-graphics software.

Examples


grSoftVersion()

[Package grDevices version 3.6.1 Index]