savePlot {grDevices}	R Documentation

Save Cairo X11 Plot to File

Description

Save the current page of a cairo X11() device to a file.

Usage

savePlot(filename = paste("Rplot", type, sep = "."),
         type = c("png", "jpeg", "tiff", "bmp"),
         device = dev.cur())

Arguments

filename
filename to save to.

type
file type: only "png" will be accepted for cairo version 1.0.

device
the device to save from.

Details

Only cairo-based X11 devices are supported.

This works by copying the image surface to a file. For PNG will always be a 24-bit per pixel PNG ‘DirectClass’ file, for JPEG the quality is 75% and for TIFF there is no compression.

For devices with buffering this copies the buffer's image surface, so works even if dev.hold has been called.

At present the plot is saved after rendering onto the canvas (default opaque white), so for the default bg = "transparent" the effective background colour is the canvas colour.

Value

Invisible NULL.

Note

There is a similar function of the same name but more types for windows devices on Windows.

See Also

X11, dev.copy, dev.print

[Package grDevices version 3.2.0 Index]