windowsFonts {grDevices}	R Documentation

Windows Fonts

Description

These functions handle the translation of a device-independent R graphics font family name to a windows font description.

Usage

windowsFont(family)

windowsFonts(...)          

Arguments

family 	 a character vector containing the font family name ("TT" as the first two characters indicates a TrueType font).

... 	 either character strings naming mappings to display, or new (named) mappings to define.

Details

A windows device is created with a default font (see the documentation for windows), but it is also possible to specify a font family when drawing to the device (for example, see the documentation for gpar in the grid package).

The font family sent to the device is a simple string name, which must be mapped to something more specific to windows fonts. A list of mappings is maintained and can be modified by the user.

The windowsFonts function can be used to list existing mappings and to define new mappings. The windowsFont function can be used to create a new mapping.

Default mappings are provided for four device-independent font family names: "sans" for a sans-serif font, "serif" for a serif font, "mono" for a monospaced font, and "symbol" for a symbol font.

These mappings will only be used if the current font face is 1 (plain), 2 (bold), 3 (italic), or 4 (bolditalic).

See Also

windows

Examples

windowsFonts()
windowsFonts("mono")

[Package grDevices version 2.1.0 Index]