xfig {grDevices}	R Documentation

XFig Graphics Device

Description

xfig starts the graphics device driver for producing XFig (version 3.2) graphics.

The auxiliary function ps.options can be used to set and view (if called without arguments) default values for the arguments to xfig and postscript.

Usage

xfig(file = ifelse(onefile, "Rplots.fig", "Rplot%03d.fig"),
     onefile = FALSE, ...)

Arguments

file 	 a character string giving the name of the file. If it is "", the output is piped to the command given by the argument command. For use with onefile = FALSE give a printf format such as "Rplot%d.fig" (the default in that case).

onefile 	 logical: if true allow multiple figures in one file. If false, assume only one page per file and generate a file number containing the page number.

... 	 further arguments to ps.options accepted by xfig():

paper

the size of paper in the printer. The choices are "A4", "Letter" and "Legal" (and these can be lowercase). A further choice is "default", which is the default. If this is selected, the papersize is taken from the option "papersize" if that is set and to "A4" if it is unset or empty.

horizontal

the orientation of the printed image, a logical. Defaults to true, that is landscape orientation.

width, height

the width and height of the graphics region in inches. The default is to use the entire page less a 0.25 inch border.

family

the font family to be used. This must be one of "AvantGarde", "Bookman", "Courier", "Helvetica", "Helvetica-Narrow", "NewCenturySchoolbook", "Palatino" or "Times".

pointsize

the default point size to be used.

bg

the default background color to be used.

fg

the default foreground color to be used.

pagecentre

logical: should the device region be centred on the page: defaults to TRUE.

Details

Although xfig can produce multiple plots in one file, the XFig format does not say how to separate or view them. So onefile = FALSE is the default.

Line widths as controlled by par(lwd=) are in multiples of 5/6*1/72 inch. Multiples less than 1 are allowed. pch="." with cex = 1 corresponds to a square of side 1/72 inch.

Note

Only some line textures (0 <= lty < 4) are used. Eventually this will be partially remedied, but the XFig file format does not allow as general line textures as the R model. Unimplemented line textures are displayed as dash-double-dotted.

There is a limit of 512 colours (plus white and black) per file.

See Also

Devices, postscript, ps.options.

[Package grDevices version 2.1.0 Index]