Devices {grDevices}	R Documentation

List of Graphical Devices

Description

The following graphics devices are currently available:

windows The graphics driver for Windows (on screen, to printer and to Windows metafile).
postscript Writes PostScript graphics commands to a file
pdf Write PDF graphics commands to a file
pictex Writes LaTeX/PicTeX graphics commands to a file
png PNG bitmap device
jpeg JPEG bitmap device
bmp BMP bitmap device
xfig Device for XFIG graphics file format
bitmap bitmap pseudo-device via GhostScript (if available).

Details

If no device is open, using a high-level graphics function will cause a device to be opened. Which device is given by options("device") which is initially set as the most appropriate for each platform: a screen device for most interactive use and postscript otherwise. The exception is interactive use under Unix if no screen device is known to be available, when postscript() is used for most systems; pdf() for Mac OS X.

See Also

The individual help files for further information on any of the devices listed here;

dev.interactive, dev.cur, dev.print, graphics.off, image, dev2bitmap.

Examples

## Not run: 
## open the default screen device on this platform if no device is
## open
if(dev.cur() == 1) get(getOption("device"))()
## End(Not run)

[Package grDevices version 2.1.0 Index]