Devices {grDevices}	R Documentation

List of Graphical Devices

Description

The following graphics devices are currently available:

windows The graphics device for Windows (on screen, to printer and to Windows metafile).

pdf Write PDF graphics commands to a file

postscript Writes PostScript graphics commands to a file

xfig Device for XFIG graphics file format

bitmap bitmap pseudo-device via Ghostscript (if available).

pictex Writes TeX/PicTeX graphics commands to a file (of historical interest only)

The following devices will be functional if R was compiled to use them (they exist but will return with a warning on other systems):

cairo_pdf, cairo_ps PDF and PostScript devices based on cairo graphics.

svg SVG device based on cairo graphics.

png PNG bitmap device

jpeg JPEG bitmap device

bmp BMP bitmap device

tiff TIFF bitmap device

Details

If no device is open, using a high-level graphics function will cause a device to be opened. Which device is given by options("device") which is initially set as the most appropriate for each platform: a screen device for most interactive use and pdf (or the setting of R_DEFAULT_DEVICE) otherwise. The exception is interactive use under Unix if no screen device is known to be available, when pdf() is used.

It is possible for an R package to provide further graphics devices and several packages on CRAN do so. These include other devices outputting SVG and PGF/TiKZ (TeX-based graphics, see http://pgf.sourceforge.net/).

See Also

The individual help files for further information on any of the devices listed here; windows.options, ps.options and pdf.options for how to customize devices.

dev.interactive, dev.cur, dev.print, graphics.off, image, dev2bitmap.

Examples

## Not run: 
## open the default screen device on this platform if no device is
## open
if(dev.cur() == 1) dev.new()

## End(Not run)

[Package grDevices version 3.5.1 Index]