bringToTop {grDevices}	R Documentation

Assign Focus to a Window

Description

bringToTop brings the specified screen device's window to the front of the window stack (and gives it focus). With first argument -1 it brings the console to the top.

If stay = TRUE, the window is designated as a topmost window, i.e. it will stay on top of any regular window. stay may only be used when Rgui is run in SDI mode. This corresponds to the “Stay on top” popup menu item in Rgui.

Usage

bringToTop(which = dev.cur(), stay = FALSE)

Arguments

which 	 a device number, or -1.

stay 	 whether to make the window stay on top.

See Also

msgWindow, windows

[Package grDevices version 2.9.2 Index]