dev.interactive {grDevices}	R Documentation

Test if an Interactive Graphics Device is in Use

Description

Test if an interactive graphics device is in use.

Usage

dev.interactive()

Details

The X11 (Unix), windows (Windows) and quartz (MacOS X) are regarded as interactive, together with GTK and gnome (used with the GNOME GUI: GTK is available in package gtkDevice and gnome is expected to be in a package gnomeDevice).

Value

dev.interactive() returns a logical, TRUE iff an interactive (screen) device is in use.

See Also

Devices for the available devices on your platform.

[Package grDevices version 2.1.0 Index]