embedFonts {grDevices}	R Documentation

Embed Fonts in PostScript and PDF

Description

Runs Ghostscript to process a PDF or PostScript file and embed all fonts in the file.

Usage

embedFonts(file, format, outfile = file, fontpaths = "",
           options = "")

Arguments

file 	 a character string giving the name of the original file.

format 	 either "pswrite" or "pdfwrite". If not specified, it is guessed from the suffix of file.

outfile 	 the name of the new file (with fonts embedded).

fontpaths 	 a character vector giving directories that Ghostscript will search for fonts.

options 	 a character string containing further options to Ghostscript.

Details

This function is not necessary if you just use the standard default fonts for PostScript and PDF output.

If you use a special font, this function is useful for embedding that font in your PostScript or PDF document so that it can be shared with others without them having to install your special font (provided the font licence allows this).

If the special font is not installed for Ghostscript, you will need to tell Ghostscript where the font is, using something like options="-sFONTPATH=path/to/font".

This function relies on a suitable Ghostscript executable being in your path, or the environment variable R_GSCMD (the same as bitmap) being set as the full path to the Ghostscript executable. For Windows users, this should be the path to gswin32c.exe. If unset, environment variable GSC is tried and then "gswin32c.exe".

Note that Ghostscript may do font substitution, so the font embedded may differ from that specified in the original file.

Value

The shell command used to invoke Ghostscript is returned invisibly. This may be useful for debugging purposes as you can run the command by hand in a shell to look for problems.

See Also

postscriptFonts, Devices.

Paul Murrell and Brian Ripley (2006) Non-standard fonts in PostScript and PDF graphics. R News, 6(2):41–47. http://cran.r-project.org/doc/Rnews/Rnews_2006-2.pdf.

[Package grDevices version 2.9.2 Index]