grDevices-package {grDevices}	R Documentation

The R Graphics Devices and Support for Colours and Fonts

Description

Graphics devices and support for base and grid graphics

Details

This package contains functions which support both base and grid graphics.

For a complete list of functions, use library(help="grDevices").

Author(s)

R Development Core Team and contributors worldwide

Maintainer: R Core Team R-core@r-project.org

[Package grDevices version 2.9.2 Index]