pictex {grDevices}	R Documentation

A PicTeX Graphics Driver

Description

This function produces graphics suitable for inclusion in TeX and LaTeX documents.

Usage

pictex(file = "Rplots.tex", width = 5, height = 4, debug = FALSE,
       bg = "white", fg = "black")

Arguments

file 	 the file where output will appear.

width 	 The width of the plot in inches.

height 	 the height of the plot in inches.

debug 	 should debugging information be printed.

bg 	 the background color for the plot. Ignored.

fg 	 the foreground color for the plot. Ignored.

Details

This driver does not have any font metric information, so the use of plotmath is not supported.

Multiple plots will be placed as separate environments in the output file.

Line widths are ignored except when setting the spacing of line textures. pch="." corresponds to a square of side 1pt.

This device does not support colour (nor does the PicTeX package), and all colour settings are ignored.

Conventions

This section describes the implementation of the conventions for graphics devices set out in the “R Internals Manual”.

The default device size is 5 inches by 5 inches.
There is no pointsize argument: the default size is interpreted as 10 point.
The only font family is cmss10.
Line widths are only used when setting the spacing on line textures.
Circle of any radius are allowed.
Colour is not supported.

Author(s)

This driver was provided by Valerio Aimale valerio@svpop.com.dist.unige.it of the Department of Internal Medicine, University of Genoa, Italy.

References

Knuth, D. E. (1984) The TeXbook. Reading, MA: Addison-Wesley.

Lamport, L. (1994) LATEX: A Document Preparation System. Reading, MA: Addison-Wesley.

Goossens, M., Mittelbach, F. and Samarin, A. (1994) The LATEX Companion. Reading, MA: Addison-Wesley.

See Also

postscript, Devices.

Examples

require(graphics)

pictex()
plot(1:11,(-5:5)^2, type='b', main="Simple Example Plot")
dev.off()
##--------------------
## Not run: 
%% LaTeX Example
\documentclass{article}
\usepackage{pictex}
\begin{document}
%...
\begin{figure}[h]
  \centerline{\input{Rplots.tex}}
  \caption{}
\end{figure}
%...
\end{document}

%%-- plain TeX Example --
\input pictex
$$ \input Rplots.tex $$
## End(Not run)
##--------------------
unlink("Rplots.tex")

[Package grDevices version 2.9.2 Index]