pictex {grDevices}	R Documentation

A PicTeX Graphics Driver

Description

This function produces simple graphics suitable for inclusion in TeX and LaTeX documents. It dates from the very early days of R and is for historical interest only.

Usage


pictex(file = "Rplots.tex", width = 5, height = 4, debug = FALSE,
       bg = "white", fg = "black")

Arguments

file
the file where output will appear.

width
The width of the plot in inches.

height
the height of the plot in inches.

debug
should debugging information be printed.

bg
the background color for the plot. Ignored.

fg
the foreground color for the plot. Ignored.

Details

This driver is much more basic than the other graphics drivers included in R. It does not have any font metric information, so the use of plotmath is not supported.

Line widths are ignored except when setting the spacing of line textures. pch = "." corresponds to a square of side 1pt.

This device does not support colour (nor does the PicTeX package), and all colour settings are ignored.

Note that text is recorded in the file as-is, so annotations involving TeX special characters (such as ampersand and underscore) need to be quoted as they would be when entering TeX.

Multiple plots will be placed as separate environments in the output file.

Conventions

This section describes the implementation of the conventions for graphics devices set out in the “R Internals Manual”.

The default device size is 5 inches by 4 inches.

There is no pointsize argument: the default size is interpreted as 10 point.

The only font family is cmss10.

Line widths are only used when setting the spacing on line textures.

Circle of any radius are allowed.

Colour is not supported.

Author(s)

This driver was provided around 1996–7 by Valerio Aimale of the Department of Internal Medicine, University of Genoa, Italy.

References

Knuth, D. E. (1984) The TeXbook. Reading, MA: Addison-Wesley.

Lamport, L. (1994) LATEX: A Document Preparation System. Reading, MA: Addison-Wesley.

Goossens, M., Mittelbach, F. and Samarin, A. (1994) The LATEX Companion. Reading, MA: Addison-Wesley.

See Also

postscript, pdf, Devices.

The tikzDevice in the CRAN package of that name for more modern TeX-based graphics (http://pgf.sourceforge.net/, although including PDF figures via pdftex is most common in (La)TeX documents).

Examples


require(graphics)


pictex()
plot(1:11, (-5:5)^2, type = "b", main = "Simple Example Plot")
dev.off()
##--------------------
## Not run: 
%% LaTeX Example
\documentclass{article}
\usepackage{pictex}
\usepackage{graphics} % for \rotatebox
\begin{document}
%...
\begin{figure}[h]
  \centerline{\input{Rplots.tex}}
  \caption{}
\end{figure}
%...
\end{document}

## End(Not run)
##--------------------
unlink("Rplots.tex")


[Package grDevices version 3.6.1 Index]