recordPlot {grDevices}	R Documentation

Record and Replay Plots

Description

Functions to save the current plot in an R variable, and to replay it.

Usage

recordPlot()
replayPlot(x)

Arguments

x 	 A saved plot.

Details

These functions record and replay the displaylist of the current graphics device. The returned object is of class "recordedplot", and replayPlot acts as a print method for that class.

Value

recordPlot returns an object of class "recordedplot".
replayPlot has no return value.

Warning

The format of recorded plots may change between R versions. Recorded plots should not be used as a permanent storage format for R plots.

R will always attempt to replay a recorded plot, but if the plot was recorded with a different R version then bad things may happen.

[Package grDevices version 2.9.2 Index]