trans3d {grDevices}	R Documentation

3D to 2D Transformation for Perspective Plots

Description

Projection of 3-dimensional to 2-dimensional points using a 4x4 viewing transformation matrix. Mainly for adding to perspective plots such as persp.

Usage

trans3d(x,y,z, pmat)

Arguments

x, y, z 	 numeric vectors of equal length, specifying points in 3D space.

pmat 	 a 4 x 4 viewing transformation matrix, suitable for projecting the 3D coordinates (x,y,z) into the 2D plane using homogeneous 4D coordinates (x,y,z,t); such matrices are returned by persp().

Value

a list with two components

x,y 	 the projected 2d coordinates of the 3d input (x,y,z).

See Also

persp

Examples

## See  help(persp) {after attaching the 'graphics' package}
##      -----------

[Package grDevices version 2.9.2 Index]