windows.options {grDevices}	R Documentation

Auxiliary Function to Set/View Defaults for Arguments of windows()

Description

The auxiliary function windows.options can be used to set or view (if called without arguments) the default values for the arguments of windows.

windows.options needs to be called before calling windows, and the default values it sets can be overridden by supplying arguments to windows.

Usage

windows.options(..., reset = FALSE)

Arguments

... 	 arguments width, height, pointsize, record, rescale, xpinch, ypinch, bg, canvas, gamma, xpos, ypos, buffered, restoreConsole, clickToConfirm and title can be supplied.

reset 	 logical: should the defaults be reset to their ‘factory-fresh’ values?

Details

If both reset = TRUE and ... are supplied the defaults are first reset to the ‘factory-fresh’ values and then the new values are applied.

Value

A named list of all the defaults. If any arguments are supplied the return values are the old values and the result has the visibility flag turned off.

See Also

windows, ps.options.

Examples

## Not run: 
## put something like this is your .Rprofile to customize the defaults
setHook(packageEvent("grDevices", "onLoad"),
        function(...)
            grDevices::windows.options(width=8, height=6, xpos=0,
                                       pointsize=10))
## End(Not run)

[Package grDevices version 2.9.2 Index]